Refactor NativeLibrariesResolver project to be more modular
