+++ tests/ComputeSharp.Tests.NativeLibrariesResolver/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+[assembly: DoNotParallelize]
